knowledge server/pyproject.toml
[project]
name = "knowledge-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "mcp>=1.22.0",
    "pymilvus>=2.6.4",
    "pypdf>=6.4.0",
    "pyyaml>=6.0.3",
    "unstructured>=0.18.21",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "ty>=0.0.1a28",
]

[tool.ty.rules]
possibly-missing-attribute = "ignore"
